refactor(chart): use `.col3-hidden` on `<body>`
Previously, `.col3-hidden` is set on elements that should change when `#col3` is
hidden, which makes toggling it a PITA.

Currently, `.col3-hidden` is set on `<body>`, and show/hide `#col3` is only a
matter of remove/add that class.

Also, the styles in this layout is moved from `footer.css` to `main.css`.